parameter if move base used in robot_manager, rename dynamic map modes
